[YottaDB/Lang/YDBGo#21] Additional Go wrapper cleanups/fixes
This change has a number of cleanups across several modules. These are mostly in the
form of removing unnecessary casts and pointer dereferences.

One somewhat larger change modifies the unexported variadicPlist struct methods to
anchor a C struct containing an array of void* pointers instead of uintptr_t integers.
Using uintptr types hides pointers from Go which ends up complicating the code
requiring extra unnecessary casts and is not "good Go" but using pointers instead
means we need a method of converting an int/uint64 to a pointer which Go does not
allow one to do so small internal C stub routines are added to do this "conversion".

The KeepAlive changes added in previous commit(s) generally take care of the issues
that the above variadicPlist changes took care of but since the change is still to
have better Go code, the change is being implemented.

There is also a potential fix in ElemLenUsed() in buffer_t_array.go where the return
statement was referencing a pointer to C memory whose anchor was protected by the
KeepAlive() on the preceding line. This was changed so the KeepAlive() is AFTER the
reference (separated from the return statement) to keep the protection intact for
that reference.

Note this change requires an accompanying change in YDB as the variadic plist call entry point
(ydb_call_variadic_plist_func()) has its argument type changed from uintptr_t to void*.

+++ util.go
@@ -138,13 +138,13 @@ func (mdesc *CallMDesc) CallMDescT(tptoken uint64, errstr *BufferT, retvallen ui
 	if nil != errstr {
 		cbuft = errstr.getCPtr()
 	}
-	err = vplist.setVPlistParam(tptoken, errstr, parmIndx, uintptr(unsafe.Pointer(cbuft)))
+	err = vplist.setVPlistParam(tptoken, errstr, parmIndx, unsafe.Pointer(cbuft))
 	if nil != err {
 		panic(fmt.Sprintf("YDB: Unknown error with varidicPlist.setVPlistParam(): %s", err))
 	}
 	parmIndx++
 	// Third parm for ydb_cip_t() is the descriptor address so add that now
-	err = vplist.setVPlistParam(tptoken, errstr, parmIndx, uintptr(unsafe.Pointer(mdesc.cmdesc.cmdesc)))
+	err = vplist.setVPlistParam(tptoken, errstr, parmIndx, unsafe.Pointer(mdesc.cmdesc.cmdesc))
 	if nil != err {
 		panic(fmt.Sprintf("YDB: Unknown error with varidicPlist.setVPlistParam(): %s", err))
 	}
@@ -156,7 +156,7 @@ func (mdesc *CallMDesc) CallMDescT(tptoken uint64, errstr *BufferT, retvallen ui
 		retvalptr.address = (*C.char)(allocMem(C.size_t(retvallen)))
 		defer freeMem(unsafe.Pointer(retvalptr.address), C.size_t(retvallen))
 		retvalptr.length = (C.ulong)(retvallen)
-		err = vplist.setVPlistParam(tptoken, errstr, parmIndx, uintptr(unsafe.Pointer(retvalptr)))
+		err = vplist.setVPlistParam(tptoken, errstr, parmIndx, unsafe.Pointer(retvalptr))
 		if nil != err {
 			return "", err
 		}
@@ -185,7 +185,7 @@ func (mdesc *CallMDesc) CallMDescT(tptoken uint64, errstr *BufferT, retvallen ui
 			defer freeMem(unsafe.Pointer(parmptr.address), C.size_t(parmptr.length))
 		}
 		// Now add parmptr to the variadic plist
-		err = vplist.setVPlistParam(tptoken, errstr, parmIndx, uintptr(unsafe.Pointer(parmptr)))
+		err = vplist.setVPlistParam(tptoken, errstr, parmIndx, unsafe.Pointer(parmptr))
 		if nil != err {
 			return "", err
 		}
